refactor(date-utils): reorganize utilities, update PersianDate, and remove obsolete files
