+++ src/index.ts
@@ -1,4 +1,3 @@
 export { default as PersianDate } from "./PersianDate";
 
-export { default as persianDate } from "./utils/createPersianDate";
-
+export { default as persianDate } from "./utils/common/createPersianDate";
